bench(m07): evaluate the corpus, and make the benchmark resist its author
Pass 1 measured the accepted M06 engine and changed nothing. Five defects
were fixed. Pass 2 re-ran all seven repositories. Route extraction moved
from 97/101 to 101/101 and model-to-table from 35/48 to 48/48, both at
precision 1.000, with every one of the 290 Queries edges confirmed by
re-reading the corpus rather than by trusting a sample.

The headline is none of that. Across seven real repositories Cartograph
produced six fully verified frontend-to-table chains and not one begins at
a frontend file. Twelve of 1,458 HttpCall edges have a TypeScript client;
the rest are Python test suites calling their own API. Every frontend in
the corpus reaches its backend through a generated client wrapper, so no
URL appears at the call site, and the cross-language claim the product
rests on went essentially unexercised. That is what M07 was for.

The evaluator scores; the runner measures; neither can reach into the
other. Recall counts only in-scope records, and the refusals, unsupported
patterns and 1,886 ambiguous decisions excluded from that denominator are
reported beside it rather than left to be inferred. What was not measured
at all — HttpCall recall, OrmAccess recall, chain recall — is stated as
plainly as what was, because four of the seven gaps predicted in advance
would surface there and their absence from the tables proves nothing.

QG-009 now enforces the artefacts: pinned commits, scope declared before
measuring, ground truth that records refusals and not only successes,
denominators equal to the in-scope record count, and results bound to the
digest of the ground truth they were scored against. Eleven attacks were
written against that gate. The first run caught ten — deleting a
ground-truth record shrinks both sides of the recall ratio and was
invisible until the digest binding closed it. All eleven are caught now.

+++ crates/cartograph-cli/tests/cli.rs
@@ -318,13 +318,13 @@ fn match_json_exports_the_graph_it_built() {
     let edges = graph["edges"].as_array().expect("edges array");
 
     assert_eq!(
-        graph["node_count"].as_u64().unwrap() as usize,
-        nodes.len(),
+        graph["node_count"].as_u64().unwrap(),
+        nodes.len() as u64,
         "the exported node list must be the whole graph"
     );
     assert_eq!(
-        graph["edge_count"].as_u64().unwrap() as usize,
-        edges.len(),
+        graph["edge_count"].as_u64().unwrap(),
+        edges.len() as u64,
         "the exported edge list must be the whole graph"
     );
     assert_eq!(
